Add module files
