+++ modules/book.js
@@ -1,8 +1,6 @@
-class Book {
+export default class Book {
   constructor(title, author) {
     this.title = title;
     this.author = author;
   }
 }
-
-export default Book;
